# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Neuralix Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guilds/{guildId}/welcome:
    get:
      operationId: getWelcomeConfig
      tags: [welcome]
      summary: Get welcome configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Welcome config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelcomeConfig"
    put:
      operationId: updateWelcomeConfig
      tags: [welcome]
      summary: Update welcome configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WelcomeConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelcomeConfig"
  /guilds/{guildId}/welcome/test:
    post:
      operationId: testWelcome
      tags: [welcome]
      summary: Send test welcome message
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test sent

  # GOODBYE
  /guilds/{guildId}/goodbye:
    get:
      operationId: getGoodbyeConfig
      tags: [goodbye]
      summary: Get goodbye configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Goodbye config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoodbyeConfig"
    put:
      operationId: updateGoodbyeConfig
      tags: [goodbye]
      summary: Update goodbye configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoodbyeConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoodbyeConfig"
  /guilds/{guildId}/goodbye/test:
    post:
      operationId: testGoodbye
      tags: [goodbye]
      summary: Send test goodbye message
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test sent

  # VERIFICATION
  /guilds/{guildId}/verification:
    get:
      operationId: getVerificationConfig
      tags: [verification]
      summary: Get verification config
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationConfig"
    put:
      operationId: updateVerificationConfig
      tags: [verification]
      summary: Update verification config
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationConfig"
  /verify/{guildId}:
    post:
      operationId: verifyUser
      tags: [verification]
      summary: Verify user for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"

  # TICKETS
  /guilds/{guildId}/tickets/config:
    get:
      operationId: getTicketConfig
      tags: [tickets]
      summary: Get ticket system configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketConfig"
    put:
      operationId: updateTicketConfig
      tags: [tickets]
      summary: Update ticket system configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketConfig"
  /guilds/{guildId}/tickets:
    get:
      operationId: getTickets
      tags: [tickets]
      summary: List tickets for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tickets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
  /guilds/{guildId}/tickets/{ticketId}/close:
    post:
      operationId: closeTicket
      tags: [tickets]
      summary: Close a ticket
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Closed

  # ANTIRAID
...
  /guilds/{guildId}/premium:
    get:
      operationId: getGuildPremium
      tags: [premium]
      summary: Get guild premium status
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Premium status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremiumStatus"
...
  /admin/licenses:
    get:
      operationId: getLicenses
      tags: [admin]
      summary: Get all licenses
      responses:
        "200":
          description: Licenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/License"
    post:
      operationId: createLicense
      tags: [admin]
      summary: Create license
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /admin/licenses/{id}:
    delete:
      operationId: revokeLicense
      tags: [admin]
      summary: Revoke license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked

  # SUPPORT
  /support/tickets:
    get:
      operationId: getSupportTickets
      tags: [support]
      summary: Get web support tickets
      responses:
        "200":
          description: Support tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createSupportTicket
      tags: [support]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /support/tickets/{id}/messages:
    get:
      operationId: getSupportMessages
      tags: [support]
      summary: Get messages for support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportMessage"
    post:
      operationId: sendSupportMessage
      tags: [support]
      summary: Send message in support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportMessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessage"

  # AI
...
components:
  schemas:
...
    WelcomeConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        embedEnabled:
          type: boolean
        embedColor:
          type: ["string", "null"]
        embedTitle:
          type: ["string", "null"]
        embedDescription:
          type: ["string", "null"]
        imageEnabled:
          type: boolean
        autoRoleIds:
          type: array
          items:
            type: string
        dmEnabled:
          type: boolean
        dmMessage:
          type: ["string", "null"]
      required: [id, guildId, enabled, embedEnabled, imageEnabled, autoRoleIds, dmEnabled]
    WelcomeConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        embedEnabled:
          type: boolean
        embedColor:
          type: ["string", "null"]
        embedTitle:
          type: ["string", "null"]
        embedDescription:
          type: ["string", "null"]
        imageEnabled:
          type: boolean
        autoRoleIds:
          type: array
          items:
            type: string
        dmEnabled:
          type: boolean
        dmMessage:
          type: ["string", "null"]
    GoodbyeConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        embedEnabled:
          type: boolean
        embedColor:
          type: ["string", "null"]
        embedTitle:
          type: ["string", "null"]
        embedDescription:
          type: ["string", "null"]
        imageEnabled:
          type: boolean
      required: [id, guildId, enabled, embedEnabled, imageEnabled]
    GoodbyeConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        embedEnabled:
          type: boolean
        embedColor:
          type: ["string", "null"]
        embedTitle:
          type: ["string", "null"]
        embedDescription:
          type: ["string", "null"]
        imageEnabled:
          type: boolean
    VerificationConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        roleId:
          type: ["string", "null"]
        minAccountAge:
          type: integer
        antiVpn:
          type: boolean
        antiAlt:
          type: boolean
        antiBot:
          type: boolean
      required: [id, guildId, enabled, minAccountAge, antiVpn, antiAlt, antiBot]
    VerificationConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        roleId:
          type: ["string", "null"]
        minAccountAge:
          type: integer
        antiVpn:
          type: boolean
        antiAlt:
          type: boolean
        antiBot:
          type: boolean
    VerificationResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        roleAssigned:
          type: boolean
      required: [success, message, roleAssigned]
    TicketConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        categoryId:
          type: ["string", "null"]
        supportRoleId:
          type: ["string", "null"]
        transcriptChannelId:
          type: ["string", "null"]
        maxTicketsPerUser:
          type: integer
        panelMessage:
          type: ["string", "null"]
        panelChannelId:
          type: ["string", "null"]
      required: [id, guildId, enabled, maxTicketsPerUser]
    TicketConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        categoryId:
          type: ["string", "null"]
        supportRoleId:
          type: ["string", "null"]
        transcriptChannelId:
          type: ["string", "null"]
        maxTicketsPerUser:
          type: integer
        panelMessage:
          type: ["string", "null"]
        panelChannelId:
          type: ["string", "null"]
    Ticket:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        channelId:
          type: string
        userId:
          type: string
        username:
          type: string
        subject:
          type: string
        status:
          type: string
        claimedBy:
          type: ["string", "null"]
        createdAt:
          type: string
        closedAt:
          type: ["string", "null"]
      required: [id, guildId, channelId, userId, username, subject, status, createdAt]
...
    PremiumStatus:
      type: object
      properties:
        guildId:
          type: string
        active:
          type: boolean
        plan:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        features:
          type: array
          items:
            type: string
      required: [guildId, active, features]
...
    License:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        plan:
          type: string
        guildId:
          type: ["string", "null"]
        active:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, key, plan, active, createdAt]
    LicenseInput:
      type: object
      properties:
        plan:
          type: string
        guildId:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
      required: [plan]
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        subject:
          type: string
        status:
          type: string
        priority:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, username, subject, status, priority, createdAt, updatedAt]
    SupportTicketInput:
      type: object
      properties:
        subject:
          type: string
        message:
          type: string
        priority:
          type: string
      required: [subject, message]
    SupportMessage:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        userId:
          type: string
        username:
          type: string
        avatar:
          type: ["string", "null"]
        content:
          type: string
        isStaff:
          type: boolean
        createdAt:
          type: string
      required: [id, ticketId, userId, username, content, isStaff, createdAt]
    SupportMessageInput:
      type: object
      properties:
        content:
          type: string
      required: [content]
...